refactor: prevent hanging execution
Instantiating a Quill instance at the module scope caused execution
of the module to hang indefinitely. Prior to Quill v1 there was a
destroy method to clean up instances however it was removed.
Removing the node in JSDOM did not resolve the issue.

We are moving instantiation into each public function. The performance
cost of doing this is unknown at this time. We will need to test at a
later time.